ELY-277 Make `terraform plan` fully determined
Conditioned `for_each` was made for avoid created list of resources (for every service, etc)
if not planned to create related resources (service-account).
But module output or dynamic datasource was used to provide conditioned variable.
This make terraform to fail since number of resources could not be determined on first run or `terraform plan`
This commt fixed it